backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
bcrypt
passlib[bcrypt]
python-dotenv
